refactor(#757): share terminal checkpoint xp parsing
Pulls the terminal-checkpoint payload parse out of getAvatarProgression
into parseTerminalCheckpointXP, so startActivity's upcoming pending-xp
scan can reuse the same rule without duplicating it.
